Model/requirements.txt
requests==2.23.0
pandas==1.0.2
numpy==1.18.1
plotly==4.5.4
bokeh==2.0.1
streamlit==0.57.3
Markdown==3.1.1
scikit_surprise==1.1.0
Pillow==7.1.1
surprise==0.1
